+++ starkbank/merchantpurchase/merchant_purchase.go
@@ -20,6 +20,7 @@ type MerchantPurchase struct {
 	HolderName         string                 `json:",omitempty"`
 	HolderEmail        string                 `json:",omitempty"`
 	HolderPhone        string                 `json:",omitempty"`
+	HolderId           string                 `json:",omitempty"`
 	FundingType        string                 `json:",omitempty"`
 	BillingCountryCode string                 `json:",omitempty"`
 	BillingCity        string                 `json:",omitempty"`